+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/tweetexample/main/domain/CarMessageBodyWriter.scala
@@ -1,6 +1,6 @@
 package com.twitter.finatra.http.tests.integration.tweetexample.main.domain
 
-import com.google.common.net.MediaType
+import com.twitter.finagle.http.MediaType
 import com.twitter.finatra.http.marshalling.{MessageBodyWriter, WriterResponse}
 import com.twitter.finatra.json.FinatraObjectMapper
 import javax.inject.Inject
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class CarMessageBodyWriter @Inject()(mapper: FinatraObjectMapper) extends MessageBodyWriter[Car] {
 
   override def write(car: Car) = {
-    WriterResponse(MediaType.JSON_UTF_8, mapper.writeValueAsBytes(Map("car" -> car.name)))
+    WriterResponse(MediaType.JsonUtf8, mapper.writeValueAsBytes(Map("car" -> car.name)))
   }
 }